HH-96502 Turn off fail fast policy when http request is not waited
Calculate 'fail_fast' for http request based on same 'fail_fast' and
'waited'. When the request is considered not waited ('waited' == False),
then we do not process request failure in appropriate method
(fail_fast branch). Also, log warning for developer that maybe such configuration is a mistake.

As a result, failed and not waited requests is never influence responses
to the client.